db_handlers/requirements.txt
apache-airflow==2.10.5
happybase==1.2.0
ipython==8.12.3
minio==7.2.15
neo4j==5.27.0
pandas==2.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pymongo==4.11.2
python-dotenv==1.0.1
yfiles_jupyter_graphs==1.9.0